requirements.txt
openai-whisper==20240930
sounddevice==0.5.1
numpy==1.26.4
#torch2trt==0.5.0
#torchaudio==2.5.1
#torchvision==0.20.1
#whisper-trt==0.0.1

#you also need to download the jetson version of torch and torchvision. See the link here:
# https://forums.developer.nvidia.com/t/pytorch-for-jetson/72048

#-f https://nvidia.box.com/shared/static/zvultzsmd4iuheykxy17s4l2n91ylpl8.whl
#-f https://nvidia.box.com/shared/static/9si945yrzesspmg9up4ys380lqxjylc3.whl
#-f https://nvidia.box.com/shared/static/u0ziu01c0kyji4zz3gxam79181nebylf.whl